refactor: make individual redux to info song modal
