+++ packages/web/src/components/SongsCard/SongsCard.jsx
@@ -3,7 +3,7 @@ import { useDispatch } from 'react-redux';
 import './styles.scss';
 
 import { play } from '../../redux/listPlayer/listPlayer-actions';
-import { openInfoModal } from '../../redux/songs/songs-actions';
+import { openInfoModal } from '../../redux/songInfoModal/songInfoModal-actions';
 import {
     useQuickMenu,
     useQuickMenuListener,
