+++ packages/web/src/redux/root-reducer.js
@@ -6,6 +6,7 @@ import uploaderReducer from './uploader/uploader-reducer';
 import songsReducer from './songs/songs-reducer';
 import quickMenu from './quickMenu/quickMenu-reducer';
 import listPlayerReducer from './listPlayer/listPlayer-reducer';
+import songInfoModal from './songInfoModal/songInfoModal-reducer';
 
 const RESET_STORE_AND_LOG_OUT = 'RESET_STORE_AND_LOG_OUT';
 
@@ -18,6 +19,7 @@ const appReducer = combineReducers({
     user: userReducer,
     uploader: uploaderReducer,
     songs: songsReducer,
+    songInfoModal,
     listPlay: listPlayerReducer,
     quickMenu,
 });
