+++ packages/web/src/redux/songInfoModal/songInfoModal-types.js
@@ -1,0 +1,2 @@
+export const OPEN_INFO_MODAL = 'OPEN_INFO_MODAL';
+export const CLOSE_INFO_MODAL = 'CLOSE_INFO_MODAL';
